project_management/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
celery==5.2.7
redis==4.5.4
django-celery-beat==2.4.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-cors-headers==4.3.1